Reorganize into folders and add favicon
